backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
openai==1.45.0
requests==2.31.0
python-dateutil==2.8.2
email-validator==2.0.0
gunicorn==21.2.0
dnspython==2.4.2
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
urllib3==2.0.5
PyJWT==2.8.0